Fix calc problem

+++ crates/control/src/led_status.rs
@@ -231,8 +231,8 @@ impl LedStatus {
 
         let minimum_temperature = 30.0;
 
-        let relative_temperature = ((current_maximum_temperature - minimum_temperature)
-            / (105.0 - minimum_temperature).floor()) as i32;
+        let relative_temperature = (current_maximum_temperature - minimum_temperature)
+            / (105.0 - minimum_temperature).floor();
 
         Ear::percentage_ears(1.0, relative_temperature)
     }
